+++ src/coreclr/debug/daccess/stdafx.h
@@ -41,6 +41,7 @@
 
 #include <common.h>
 #include <codeman.h>
+#include <dbgipcevent_runtimeside.h>
 #include <debugger.h>
 #include <walker.h>
 #include <controller.h>
