+++ src/coreclr/debug/di/rspriv.h
@@ -57,392 +57,9 @@ struct MachineInfo;
 #define CRASH(x)  _ASSERTE(!(x))
 #define ASSERT(x) _ASSERTE(x)
 
-//
-// Event structure that is passed between the Runtime Controller and the
-// Debugger Interface. Some types of events are a fixed size and have
-// entries in the main union, while others are variable length and have
-// more specialized data structures that are attached to the end of this
-// structure.
-//
-struct MSLAYOUT DebuggerIPCEvent_DebuggerSide
-{
-    DebuggerIPCEventType    type;
-    DWORD             processId;
-    DWORD             threadId;
-    VMPTR_AppDomain   vmAppDomain;
-    VMPTR_Thread      vmThread;
 
-    HRESULT           hr;
-    bool              replyRequired;
-    bool              asyncSend;
+#include <dbgipcevent_debuggerside.h>
 
-    union MSLAYOUT
-    {
-        struct MSLAYOUT
-        {
-            // Module whose metadata is being updated
-            // This tells the RS that the metadata for that module has become invalid.
-            VMPTR_Assembly vmAssembly;
-
-        } MetadataUpdateData;
-
-        struct MSLAYOUT
-        {
-            // Handle to CLR's internal appdomain object.
-            VMPTR_AppDomain vmAppDomain;
-        } AppDomainData;
-
-        struct MSLAYOUT
-        {
-            VMPTR_Assembly vmAssembly;
-        } AssemblyData;
-
-        // Debug event that a module has been loaded
-        struct MSLAYOUT
-        {
-            // Module that was just loaded.
-            VMPTR_Assembly vmAssembly;
-        }LoadModuleData;
-
-
-        struct MSLAYOUT
-        {
-            VMPTR_Assembly vmAssembly;
-        } UnloadModuleData;
-
-
-        // The given module's pdb has been updated.
-        // Query PDB from OOP
-        struct MSLAYOUT
-        {
-            VMPTR_Assembly vmAssembly;
-        } UpdateModuleSymsData;
-
-        struct MSLAYOUT
-        {
-            LSPTR_BREAKPOINT breakpointToken;
-            mdMethodDef  funcMetadataToken;
-            VMPTR_Assembly vmAssembly;
-            bool         isIL;
-            SIZE_T       offset;
-            SIZE_T       encVersion;
-            LSPTR_METHODDESC  nativeCodeMethodDescToken; // points to the MethodDesc if !isIL
-            CORDB_ADDRESS codeStartAddress;
-        } BreakpointData;
-
-        struct MSLAYOUT
-        {
-            mdMethodDef funcMetadataToken;
-            VMPTR_Module pModule;
-        } DisableOptData;
-
-        struct MSLAYOUT
-        {
-            BOOL enableEvents;
-            VMPTR_Object vmObj;
-        } ForceCatchHandlerFoundData;
-
-        struct MSLAYOUT
-        {
-            VMPTR_Module vmModule;
-            mdTypeDef    classMetadataToken;
-            BOOL Enabled;
-        } CustomNotificationData;
-
-        struct MSLAYOUT
-        {
-            LSPTR_BREAKPOINT breakpointToken;
-        } BreakpointSetErrorData;
-
-        struct MSLAYOUT
-        {
-#ifdef FEATURE_DATABREAKPOINT
-            CONTEXT context;
-#else
-            int dummy;
-#endif
-        } DataBreakpointData;
-
-        struct MSLAYOUT
-        {
-            LSPTR_STEPPER        stepperToken;
-            VMPTR_Thread         vmThreadToken;
-            FramePointer         frameToken;
-            bool                 stepIn;
-            bool                 rangeIL;
-            bool                 IsJMCStop;
-            unsigned int         totalRangeCount;
-            CorDebugStepReason   reason;
-            CorDebugUnmappedStop rgfMappingStop;
-            CorDebugIntercept    rgfInterceptStop;
-            unsigned int         rangeCount;
-            COR_DEBUG_STEP_RANGE range[5]; //note that this is an array
-        } StepData;
-
-        // Allocate memory on the left-side
-        struct MSLAYOUT
-        {
-            ULONG      bufSize;             // number of bytes to allocate
-        } GetBuffer;
-
-        // Memory allocated on the left-side
-        struct MSLAYOUT
-        {
-            void        *pBuffer;           // LS pointer to the buffer allocated
-            HRESULT     hr;                 // success / failure
-        } GetBufferResult;
-
-        // Free a buffer allocated on the left-side with GetBuffer
-        struct MSLAYOUT
-        {
-            void        *pBuffer;           // Pointer previously returned in GetBufferResult
-        } ReleaseBuffer;
-
-        struct MSLAYOUT
-        {
-            HRESULT     hr;
-        } ReleaseBufferResult;
-
-        // Apply an EnC edit
-        struct MSLAYOUT
-        {
-            VMPTR_Assembly vmAssembly;      // Module to edit
-            DWORD cbDeltaMetadata;              // size of blob pointed to by pDeltaMetadata
-            CORDB_ADDRESS pDeltaMetadata;       // pointer to delta metadata in debuggee
-                                                // it's the RS's responsibility to allocate and free
-                                                // this (and pDeltaIL) using GetBuffer / ReleaseBuffer
-            CORDB_ADDRESS pDeltaIL;             // pointer to delta IL in debugee
-            DWORD cbDeltaIL;                    // size of blob pointed to by pDeltaIL
-        } ApplyChanges;
-
-        struct MSLAYOUT
-        {
-            HRESULT hr;
-        } ApplyChangesResult;
-
-        struct MSLAYOUT
-        {
-            mdTypeDef   classMetadataToken;
-            VMPTR_Assembly vmAssembly;
-        } LoadClass;
-
-        struct MSLAYOUT
-        {
-            mdTypeDef   classMetadataToken;
-            VMPTR_Assembly vmAssembly;
-        } UnloadClass;
-
-        struct MSLAYOUT
-        {
-            VMPTR_Assembly vmAssembly;
-            bool  flag;
-        } SetClassLoad;
-
-        struct MSLAYOUT
-        {
-            VMPTR_OBJECTHANDLE vmExceptionHandle;
-            bool        firstChance;
-            bool        continuable;
-        } Exception;
-
-        struct MSLAYOUT
-        {
-            void        *address;
-        } IsTransitionStub;
-
-        struct MSLAYOUT
-        {
-            bool        isStub;
-        } IsTransitionStubResult;
-
-        struct MSLAYOUT
-        {
-            CORDB_ADDRESS    startAddress;
-            bool             fCanSetIPOnly;
-            VMPTR_Thread     vmThreadToken;
-            VMPTR_Assembly vmAssembly;
-            mdMethodDef      mdMethod;
-            VMPTR_MethodDesc vmMethodDesc;
-            SIZE_T           offset;
-            bool             fIsIL;
-        } SetIP; // this is also used for CanSetIP
-
-        struct MSLAYOUT
-        {
-            int iLevel;
-
-            CORDB_ADDRESS szCategory;
-            CORDB_ADDRESS szContent;
-        } FirstLogMessage;
-
-        struct MSLAYOUT
-        {
-            int iLevel;
-        } LogSwitchSettingMessage;
-
-        // information needed to send to the RS as part of a custom notification from the target
-        struct MSLAYOUT
-        {
-            // assembly for the domain in which the notification occurred
-            VMPTR_Assembly vmAssembly;
-
-            // metadata token for the type of the CustomNotification object's type
-            mdTypeDef    classToken;
-        } CustomNotification;
-
-        struct MSLAYOUT
-        {
-            VMPTR_Thread vmThreadToken;
-            CorDebugThreadState debugState;
-        } SetAllDebugState;
-
-        DebuggerIPCE_FuncEvalInfo FuncEval;
-
-        struct MSLAYOUT
-        {
-            CORDB_ADDRESS argDataArea;
-            LSPTR_DEBUGGEREVAL debuggerEvalKey;
-        } FuncEvalSetupComplete;
-
-        struct MSLAYOUT
-        {
-            RSPTR_CORDBEVAL funcEvalKey;
-            bool            successful;
-            bool            aborted;
-            void           *resultAddr;
-
-            // AppDomain that the result is in.
-            VMPTR_AppDomain vmAppDomain;
-
-            VMPTR_OBJECTHANDLE vmObjectHandle;
-            ExpandedTypeData_DebuggerSide resultType;
-        } FuncEvalComplete;
-
-        struct MSLAYOUT
-        {
-            LSPTR_DEBUGGEREVAL debuggerEvalKey;
-        } FuncEvalAbort;
-
-        struct MSLAYOUT
-        {
-            LSPTR_DEBUGGEREVAL debuggerEvalKey;
-        } FuncEvalRudeAbort;
-
-        struct MSLAYOUT
-        {
-            LSPTR_DEBUGGEREVAL debuggerEvalKey;
-        } FuncEvalCleanup;
-
-        struct MSLAYOUT
-        {
-            void           *objectRefAddress;
-            VMPTR_OBJECTHANDLE vmObjectHandle;
-            void           *newReference;
-        } SetReference;
-
-        struct MSLAYOUT
-        {
-            NameChangeType  eventType;
-            VMPTR_AppDomain vmAppDomain;
-            VMPTR_Thread    vmThread;
-        } NameChange;
-
-        // EnC Remap opportunity
-        struct MSLAYOUT
-        {
-            VMPTR_Assembly vmAssembly;
-            mdMethodDef funcMetadataToken ;        // methodDef of function with remap opportunity
-            SIZE_T          currentVersionNumber;  // version currently executing
-            SIZE_T          resumeVersionNumber;   // latest version
-            SIZE_T          currentILOffset;       // the IL offset of the current IP
-            SIZE_T          *resumeILOffset;       // pointer into left-side where an offset to resume
-                                                   // to should be written if remap is desired.
-        } EnCRemap;
-
-        // EnC Remap has taken place
-        struct MSLAYOUT
-        {
-            VMPTR_Assembly vmAssembly;
-            mdMethodDef funcMetadataToken;         // methodDef of function that was remapped
-        } EnCRemapComplete;
-
-        // Notification that the LS is about to update a CLR data structure to account for a
-        // specific edit made by EnC (function add/update or field add).
-        struct MSLAYOUT
-        {
-            VMPTR_Assembly vmAssembly;
-            mdToken         memberMetadataToken;   // Either a methodDef token indicating the function that
-                                                   // was updated/added, or a fieldDef token indicating the
-                                                   // field which was added.
-            mdTypeDef       classMetadataToken;    // TypeDef token of the class in which the update was made
-            SIZE_T          newVersionNumber;      // The new function/module version
-        } EnCUpdate;
-
-        struct MSLAYOUT
-        {
-            void      *oldData;
-            void      *newData;
-            BasicTypeData_DebuggerSide type;
-        } SetValueClass;
-
-
-        // Event used to tell LS if a single function is user or non-user code.
-        // Same structure used to get function status.
-        // @todo - Perhaps we can bundle these up so we can set multiple funcs w/ 1 event?
-        struct MSLAYOUT
-        {
-            VMPTR_Assembly vmAssembly;
-            mdMethodDef     funcMetadataToken;
-            DWORD           dwStatus;
-        } SetJMCFunctionStatus;
-
-        struct MSLAYOUT
-        {
-            void               *objectToken;
-            CorDebugHandleType handleType;
-        } CreateHandle;
-
-        struct MSLAYOUT
-        {
-            VMPTR_OBJECTHANDLE vmObjectHandle;
-        } CreateHandleResult;
-
-        // used in DB_IPCE_DISPOSE_HANDLE event
-        struct MSLAYOUT
-        {
-            VMPTR_OBJECTHANDLE vmObjectHandle;
-            CorDebugHandleType handleType;
-        } DisposeHandle;
-
-        struct MSLAYOUT
-        {
-            FramePointer                  framePointer;
-            SIZE_T                        nOffset;
-            CorDebugExceptionCallbackType eventType;
-            DWORD                         dwFlags;
-            VMPTR_OBJECTHANDLE            vmExceptionHandle;
-        } ExceptionCallback2;
-
-        struct MSLAYOUT
-        {
-            CorDebugExceptionUnwindCallbackType eventType;
-            DWORD                               dwFlags;
-        } ExceptionUnwind;
-
-        struct MSLAYOUT
-        {
-            VMPTR_Thread vmThreadToken;
-            FramePointer frameToken;
-        } InterceptException;
-
-        struct MSLAYOUT
-        {
-            VMPTR_Module vmModule;
-            void * pMetadataStart;
-            ULONG nMetadataSize;
-        } MetadataUpdateRequest;
-    };
-};
 
 // We want to keep the 'worst' HRESULT - if one has failed (..._E_...) & the
 // other hasn't, take the failing one.  If they've both/neither failed, then
@@ -5142,12 +4759,12 @@ public:
     // type data is requested but was not fetched correctly during Init()
     HRESULT TypeToBasicTypeData(BasicTypeData_DebuggerSide *data);
     void TypeToExpandedTypeData(ExpandedTypeData_DebuggerSide *data);
-    void TypeToTypeArgData(DebuggerIPCE_TypeArgData *data);
+    void TypeToTypeArgData(DebuggerIPCE_TypeArgData_DebuggerSide *data);
 
     void CountTypeDataNodes(unsigned int *count);
     static void CountTypeDataNodesForInstantiation(unsigned int genericArgsCount, ICorDebugType *genericArgs[], unsigned int *count);
-    static void GatherTypeData(CordbType *type, DebuggerIPCE_TypeArgData **curr_tyargData);
-    static void GatherTypeDataForInstantiation(unsigned int genericArgsCount, ICorDebugType *genericArgs[], DebuggerIPCE_TypeArgData **curr_tyargData);
+    static void GatherTypeData(CordbType *type, DebuggerIPCE_TypeArgData_DebuggerSide **curr_tyargData);
+    static void GatherTypeDataForInstantiation(unsigned int genericArgsCount, ICorDebugType *genericArgs[], DebuggerIPCE_TypeArgData_DebuggerSide **curr_tyargData);
 
     HRESULT GetParentType(CordbClass * baseClass, CordbType ** ppRes);
 
@@ -10296,7 +9913,7 @@ public:
     // Non-COM methods
     //-----------------------------------------------------------
     HRESULT GatherArgInfo(ICorDebugValue *pValue,
-                          DebuggerIPCE_FuncEvalArgData *argData);
+                          DebuggerIPCE_FuncEvalArgData_DebuggerSide *argData);
     HRESULT SendCleanup();
 
     // Create a RS literal for primitive type funceval result. In case the result is used as an argument for
